fixed parsers, fix ovitem, add warp to anomaly endpoint, note: need to fix parse location info

+++ Domen/Models/OverviewItem.cs
@@ -22,6 +22,6 @@ namespace Domen.Models
         public Distance Distance { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
-        public int Speed { get; set; }
+        public long Speed { get; set; }
     }
 }
